+++ podcasts/PlaybackQueue.swift
@@ -321,8 +321,7 @@ class PlaybackQueue: NSObject {
 
         DispatchQueue.global().async { [weak self] in
             guard let self = self else { return }
-
-            let episodes = self.allEpisodes()
+            let episodes = self.allEpisodes(includeNowPlaying: !FeatureFlag.cachePlayingEpisode.enabled)
             for episode in episodes {
                 self.autoDownloadIfRequired(episode: episode)
             }